# PERF: Groupby.any/all object dtype