```
from asv_bench.benchmarks.groupby import *
self = GroupByMethods()
self.setup("object", "any", "direct", 10)
 

%timeit self.time_dtype_as_field("object", "any", "direct", 10)
1.12 ms ± 60.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
11.4 ms ± 49.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   # <- master
```